backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
supabase
pydantic-settings
httpx
python-dotenv
aiohttp
